eda_cryptocurrency_app/requirements.txt
streamlit==1.9.2
Pillow==9.2.0
pandas==1.5.1
numpy==1.23.4
pybase64==1.2.3
matplotlib==3.6.0
requests==2.28.1
